tensor model initial cleanup
drop Mask concept
move implementation to cpp file
