--- eval/src/vespa/eval/eval/test/tensor_model.hpp
@@ -1,77 +1,0 @@
-// Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-
-#pragma once
-
-#include "tensor_model.h"
-
-namespace vespalib::eval::test {
-
-Layout float_cells(const Layout &layout) {
-    return Layout(CellType::FLOAT, layout.domains);
-}
-
-Domain x() { return Domain("x", {}); }
-Domain x(size_t size) { return Domain("x", size); }
-Domain x(const std::vector<vespalib::string> &keys) { return Domain("x", keys); }
-
-Domain y() { return Domain("y", {}); }
-Domain y(size_t size) { return Domain("y", size); }
-Domain y(const std::vector<vespalib::string> &keys) { return Domain("y", keys); }
-
-Domain z() { return Domain("z", {}); }
-Domain z(size_t size) { return Domain("z", size); }
-Domain z(const std::vector<vespalib::string> &keys) { return Domain("z", keys); }
-
-// Infer the tensor type spanned by the given spaces
-vespalib::string infer_type(const Layout &layout) {
-    std::vector<ValueType::Dimension> dimensions;
-    for (const auto &domain: layout) {
-        if (domain.size == 0) {
-            dimensions.emplace_back(domain.dimension); // mapped
-        } else {
-            dimensions.emplace_back(domain.dimension, domain.size); // indexed
-        }
-    }
-    return ValueType::tensor_type(dimensions, layout.cell_type).to_spec();
-}
-
-TensorSpec spec(const Layout &layout, const Sequence &seq, const Mask &mask) {
-    return TensorSpecBuilder(layout, seq, mask).build();
-}
-TensorSpec spec(const Layout &layout, const Sequence &seq) {
-    return spec(layout, seq, All());
-}
-TensorSpec spec(const Layout &layout) {
-    return spec(layout, Seq(), None());
-}
-TensorSpec spec(const Domain &domain, const Sequence &seq, const Mask &mask) {
-    return spec(Layout({domain}), seq, mask);
-}
-TensorSpec spec(const Domain &domain, const Sequence &seq) {
-    return spec(Layout({domain}), seq);
-}
-TensorSpec spec(const Domain &domain) {
-    return spec(Layout({domain}));
-}
-TensorSpec spec(double value) {
-    return spec(Layout({}), Seq({value}));
-}
-TensorSpec spec() {
-    return spec(Layout({}));
-}
-
-TensorSpec spec(const vespalib::string &type,
-                const std::vector<std::pair<TensorSpec::Address, TensorSpec::Value>> &cells) {
-    TensorSpec spec("tensor(" + type + ")");
-
-    for (const auto &cell : cells) {
-        spec.add(cell.first, cell.second);
-    }
-    return spec;
-}
-
-TensorSpec spec(const vespalib::string &value_expr) {
-    return TensorSpec::from_expr(value_expr);
-}
-
-}
